requirements.txt
# Core dependencies for EMTP (Expert Model Training Pipeline)

# URL retrieval dependencies
ddgs>=0.1.0

# Screenshot capture dependencies
selenium>=4.15.0
Pillow>=10.0.0
webdriver-manager>=4.0.0
pydoll-python>=2.9.0

# Screenshot processing dependencies
pytesseract>=0.3.10
language-tool-python>=2.9.4

# PDF processing dependencies
pypdf>=4.0.0

requests

pyyaml